+++ vms-ide/src/synchronizer/config/sections/synchronize.ts
@@ -10,7 +10,6 @@ export class SynchronizeSection implements ISynchronizeSection, IConfigSection {
         typeof candidate.downloadNewFiles === "string" &&
         typeof candidate.keepAlive === "boolean" &&
         typeof candidate.setTimeAttempts === "number" &&
-        typeof candidate.setTimeByShell === "boolean" &&
         typeof candidate.unzipCmd === "string" &&
         typeof candidate.preferZip === "boolean" &&
         typeof candidate.forceLocalTime === "boolean" &&
@@ -20,7 +19,6 @@ export class SynchronizeSection implements ISynchronizeSection, IConfigSection {
     public downloadNewFiles: DownloadAction = "edit";
     public keepAlive: boolean = false;
     public setTimeAttempts: number = 3;
-    public setTimeByShell: boolean = false;
     public unzipCmd: string = "";
     public preferZip: boolean = false;
     public forceLocalTime: boolean = true;
@@ -42,7 +40,6 @@ export class SynchronizeSection implements ISynchronizeSection, IConfigSection {
             forceLocalTime: this.forceLocalTime,
             purge: this.purge,
             setTimeAttempts: this.setTimeAttempts,
-            setTimeByShell: this.setTimeByShell,
             unzipCmd: this.unzipCmd,
         };
     }
@@ -61,7 +58,6 @@ export class SynchronizeSection implements ISynchronizeSection, IConfigSection {
             }
             this.keepAlive = data.keepAlive;
             this.setTimeAttempts = data.setTimeAttempts;
-            this.setTimeByShell = data.setTimeByShell;
             this.unzipCmd = data.unzipCmd;
             this.preferZip = data.preferZip;
             this.forceLocalTime = data.forceLocalTime;